# Improve efficiency of casting complex numbers to complex vectors.