@heltonmc 2023-04-17
After some benchmarks this does improve performance slightly and greatly reduces the llvm code generated and native code generation

```julia

function test(x, y)
    @assert length(x) == length(y)
    out = SIMDMath.ComplexVec{4, Float64}((0.0, 0.0, 0.0, 0.0), (0.0, 0.0, 0.0, 0.0))
    for i in eachindex(x)
        out = SIMDMath.fadd(out, SIMDMath.fmul(SIMDMath.ComplexVec(x[i]), SIMDMath.ComplexVec(y[i])))
    end
    return out
end
function test2(x, y)
    @assert length(x) == length(y)
    out = (0.0 + 0.0im, 0.0 + 0.0im, 0.0 + 0.0im, 0.0 + 0.0im)
    for i in eachindex(x)
        out = @. out + x[i] * y[i]
    end
    return out
end
```

```julia
# base (auto-vectorizer)
julia> @benchmark test2($x, $y)
BenchmarkTools.Trial: 10000 samples with 955 evaluations.
 Range (min … max):  92.932 ns … 137.261 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     93.020 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   93.420 ns ±   1.941 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▅█▄▃                              ▁▃▁   ▁                    ▁
  ████▆▃▄▃▃▃▁▃▁▅▁▄▁▄▄▃▄▃▃▄▄▅▃▃▁▃▁▃▁▆███▅▅███▆▅▃▄▁▄▅▃▅▃▅▆▆▅▆▅▅▆ █
  92.9 ns       Histogram: log(frequency) by time      97.9 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

# master branch
julia> @benchmark test($x, $y)
BenchmarkTools.Trial: 10000 samples with 969 evaluations.
 Range (min … max):  80.753 ns … 116.185 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     80.840 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   81.129 ns ±   1.557 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▅█▅▂                                 ▁▃                      ▁
  ████▆▅▄▄▁▁▃▁▄▃▁▃▄▄▃▃▅▄▄▃▁▃▁▁▁▁▃▁▁▁▁▆███▅▆▄▄▄▁▁▁▄▁▅▃▁▅▅▅▄▅▆▆▅ █
  80.8 ns       Histogram: log(frequency) by time      85.4 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

# this PR
julia> @benchmark test($x, $y)
BenchmarkTools.Trial: 10000 samples with 970 evaluations.
 Range (min … max):  77.620 ns … 114.777 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     77.749 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   78.005 ns ±   1.677 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▄█▆▃▂                                 ▂▂                     ▁
  █████▆▄▄▃▃▃▃▄▃▃▁▄▅▃▁▄▃▄▃▄▄▁▁▁▁▃▄▃▃▄▁▃███▇▅▆▁▄▄▄▁▃▁▃▁▄▄▄▅▄▅▆▆ █
  77.6 ns       Histogram: log(frequency) by time      82.1 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```
This is a clear improvement over the current method so will merge. The discussion in the other thread can be left open as the discussion on which format to use is separate from improving the implementation of one method.